Move methods out of CorePivotTable
-  Couple of methods are really small utility functions on top of
   ag-grid API. Moved those out to agGridUtils.ts

JIRA: FET-608
